#72 - refactor : Twitch StreamData 관련 코드 레포지토리 분리
