+++ core/data/src/main/java/com/hegunhee/data/repository/DefaultRepository.kt
@@ -29,41 +29,6 @@ class DefaultRepository @Inject constructor(
     private val localDataSource : LocalDataSource
 ) : Repository {
 
-    override suspend fun getStreamData(streamerId: String): Result<StreamDataType> {
-        return runCatching {
-            val streamerInfo = remoteDataSource
-                .getStreamerDataResponse(streamerId = arrayOf(streamerId))
-                .findByStreamerId(streamerId)
-
-            val streamData = remoteDataSource.getStreamDataResponse(streamerId = streamerInfo.streamerId)
-
-            if (streamData.isEmpty()) {
-                streamerInfo.toOfflineData()
-            } else {
-                streamData.streamApiData[0].toStreamData(streamerInfo.profileImageUrl)
-            }
-        }
-    }
-
-    override suspend fun getStreamDataList(): Result<List<StreamDataType>> {
-        return runCatching {
-            val loadedStreamerLoginArray = localDataSource.getAllStreamerList().map { it.streamerLogin }.toTypedArray()
-
-            if (loadedStreamerLoginArray.isEmpty()) {
-                return@runCatching emptyList<StreamDataType>()
-            }
-
-            val streamerInfoList = remoteDataSource.getStreamerDataResponse(streamerId = loadedStreamerLoginArray).streamerApiDataList
-            val streamInfoList =
-                remoteDataSource.getStreamDataListResponse(streamerId = loadedStreamerLoginArray).streamApiData
-
-            streamerInfoList.map { streamerInfo ->
-                val streamDataOrNull = streamInfoList.find { it.streamerId == streamerInfo.streamerId }
-                return@map streamDataOrNull?.toStreamData(streamerInfo.profileImageUrl) ?: streamerInfo.toOfflineData()
-            }
-        }
-    }
-
     override suspend fun getGameStreamDataList(gameId: String): Result<List<StreamDataType.OnlineData>> {
         return runCatching {
             val gameStreamList = remoteDataSource.getGameStreamDataResponse(gameId).streamApiData.sortedBy { it.streamerId }
