+++ core/data/src/test/java/com/hegunhee/data/repository/twitch/StreamDataTest.kt
@@ -24,7 +24,7 @@ import org.mockito.kotlin.whenever
 class StreamDataTest {
 
     @InjectMocks
-    private lateinit var sut: DefaultRepository
+    private lateinit var sut: DefaultTwitchStreamRepository
 
     @Mock
     private lateinit var remoteDataSource: RemoteDataSource
