+++ core/domain/src/main/java/com/hegunhee/domain/repository/twitch/TwitchStreamRepository.kt
@@ -1,0 +1,11 @@
+package com.hegunhee.domain.repository.twitch
+
+import com.hegunhee.domain.model.twitch.StreamDataType
+
+interface TwitchStreamRepository {
+
+    suspend fun getStreamData(streamerId: String): Result<StreamDataType>
+
+    suspend fun getStreamDataList(): Result<List<StreamDataType>>
+
+}
