+++ core/domain/src/main/java/com/hegunhee/domain/usecase/twitch/GetBookmarkedStreamDataListUseCase.kt
@@ -1,12 +1,12 @@
 package com.hegunhee.domain.usecase.twitch
 
 import com.hegunhee.domain.model.twitch.StreamDataType
-import com.hegunhee.domain.repository.Repository
+import com.hegunhee.domain.repository.twitch.TwitchStreamRepository
 import javax.inject.Inject
 
-class GetBookmarkedStreamDataListUseCase @Inject constructor(private val repository: Repository) {
+class GetBookmarkedStreamDataListUseCase @Inject constructor(private val twitchStreamRepository: TwitchStreamRepository) {
 
     suspend operator fun invoke() : Result<List<StreamDataType>>{
-        return repository.getStreamDataList()
+        return twitchStreamRepository.getStreamDataList()
     }
 }
