+++ core/domain/src/main/java/com/hegunhee/domain/usecase/twitch/GetStreamDataUseCase.kt
@@ -1,12 +1,12 @@
 package com.hegunhee.domain.usecase.twitch
 
 import com.hegunhee.domain.model.twitch.StreamDataType
-import com.hegunhee.domain.repository.Repository
+import com.hegunhee.domain.repository.twitch.TwitchStreamRepository
 import javax.inject.Inject
 
-class GetStreamDataUseCase @Inject constructor(private val repository: Repository) {
+class GetStreamDataUseCase @Inject constructor(private val twitchStreamRepository: TwitchStreamRepository) {
 
-    suspend operator fun invoke(streamerId : String) : Result<StreamDataType>{
-        return repository.getStreamData(streamerId)
+    suspend operator fun invoke(streamerId: String): Result<StreamDataType> {
+        return twitchStreamRepository.getStreamData(streamerId)
     }
 }
